[SW-100] Trace Parser PDTrace support
Committing the early scripts that we have been using for trace
decoding. This was made mainly with the output traces coming
from the simulation environment, and will now need to be extended
for the more detailed information available in PDTrace.

This commit gives us basic execution trace of the program that
we're running, using the trace file (sim or PDTrace DQ'd in TM
format).