Get the wu_trace utility running again.
Simplify how we build the WU hierarchy.  Previously, the code tried
to infer which transactions were tightly related and which were loosely
related.  Now, WUs are divided between
transactions - triggered by a new hardware event, or because of a
WU_TRANSACTION_START menu.  Each transaction has a root event with
successors.  Successors are the WUs that are explicitly run because of the
sender.

For the HTML view, each transaction gets its own time line (though we
can change it to allow chains of related WUs to be grouped in the
timeline.)  For Graphviz, transactions with the same starting WU are
grouped together.

Graphviz results are now more readable; common WUs are duplicated
in each transaction.

There's still a bunch of work to do on wu_trace, but this gives us a
running base to start from.