SSE interface first draft
There are some challenges with this interface, due to me wanting
to include it in the ADMIN model (CREATE, DESTROY, ...). If we want
to do a WRITE, the type of entries that we would want to do can
vary a lot in size, requiring another UNION. This creates unions
within unions, which starts getting messy.