Move bank and index to per-entry fields
Previously, the assumption was that all entries provided in an FFE
request apply to the same stage, table, and bank, and are in consecutive
indexes.

This was too restrictive, so we moved the index and bank, assuming that
a batch would still apply to a specific stage+table, but allowing
for non-consecutive entries and banks.

This gives the host app control to batch entries, including double word
accesses.