Save accurate cycle count in fundata output of tprs
In PDTrace outputs, we are observing two pipelines run in parallel, and
each entry we see increments the cycle count by 1. This means that we're
incrementing twice as fast as is really the case, and we're reporting
too many cycles for operations.

Flow:
pipe0: [0: instr][2: instr][4: instr][...]
pipe1: [1: instr][3: instr][5: instr][...]

The above would previously report 6 cycles, when there were only actually 3.

Open question: Currently assuming single pipeline for sim and qemu outputs,
to be adjusted if needed.