Change wu_trace to handle current output when using --wulog.
FunOS PR#820 changes the format of the WU log output by removing
the timer trigger events.  In the past, we got events for the timer trigger
and the queued wu send events.  This won't work on the real hardware
because software doesn't get control.  Instead, wu_trace now expects
to see a TIMER START event, followed by a WU START event for the same wu,
VP, and arg0.