Change header template to generate declarations in the same order they were seen in the .gen file.
Replace declaration_kind with individual flags for each kind of declaration
to make it obvious how the kind test works in the template.

Use macros in the Jinja templates to move the template code for enums,
flagsets, and structs out of the loop.

Pass in the list of raw declarations (in order) as a new global variable
in templates.