Add support for generating json in parallel with the existing picked output.
The JSON dictionary in the output file (fundata_c?.json) has three bits:

call_trees: a list of all nodes in the call tree.  The calls field uses numbered
indexes to identify nesting.
roots: a list of the root nodes in the call tree for each VP.  Used to identify
starting points in the call trees.
function_stats: a dictionary of function name -> dictionary giving summary
details on each function such as the number of calls; min, max, and average
latency for each, and cache miss information.

For now, this is a proof of concept to show we can load this data into
the Palladium Jobs web server.  We also output the JSON before we attempt
to output the pickled data to avoid crashes during pickling.