removed hack from fun_json. decode incoming object according to its type. fun_commander intended to use the array argument to contain actual data so will take it as is. array has strict order, so user better place the right argument at the right place
if the 1st argument is a dictionary, we assume the user put all the arguments
in the dictionary so only 1 argument is needed. In this case, ordering doesn't
matter. but the key must match the fields in the struct.

added struct field support.
bit field is a bit of a hack.
union is not supported.