generator.py: fix indent support
Commit 384b91b4 ('Check for errors on indent failure and fall back to
non-indent path') introduced a check for failing indent invocations.
However, that leaves the generated files unindented.

The indentation failure is usually caused by an incompatible set of
indent flags, which may vary depending on the version of indent tool
used. For example '-nfcb' is not supported in a typical linux distro
variant.

Run a check of available options and only select the supported ones
before running indent to produce output as close as possible to the
expected one.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>