Add a new sort of call instruction to find when parsing the disassembly. This didn't appear in FunOS, but does appear in UBoot.
Background: The Samurai traces show sequences of instructions.  They don't
include anything on what the instructions are or how they affect the stack -
which are calls, jump, branches, or returns.  Instead, we parse the assembly,
note the addresses where control flow changes, and then use those to
reconstruct the call graph afterwards.

The bal instruction is a branch-and-link - a PC relative call.  It's like
jal (jump-and-link) but uses a PC relative address but otherwise behaves
the same.  It's treated differently than balc (branch-and-link-compact)
because the compact instructions don't execute the instruction after the
jump before actually going to the new address.