Added a hack so that "help csr" prints something useful.
This is a pretty gross hack because "help foo" gets executed on the FunOS side, normally, but in the case of "help csr", because csr is a macro applied dpcsh-side, the help should come from the commander on the client side.

I can think of several solutions for solving this cleanly, but all involve many more lines of code than the hack in this pr, and since this is just to help debug, I dont think it's worth it at this time.