SWTOOLS-552 Change pcieproxy.c WRITE command to include explicit register size.
Instead of intuiting the Register Size as 64-bits * the number of values
provided on the WRITE command line, we now have the WRITE command take the
Register Size (in bits) explicitly.  This isn't technically needed since the
Control/Status Register Control Unit Command deals in units of 64-bits as
far as the Register Sizes, but it makes things a bit cleaner and more
explicit.