SWTOOLS-552: Get the PCIe Proxy working on FS1600 F1_1
Several interesting issues were uncovered in getting things to work on the
second F1 of the FS1600.  Bugs in FunHW, FunOS, and bad assumptions in
FunTools.

For FunTools the big issue was were were assuming that the Control/Status
Register Control Unit Spinlock configured would be the first one.  But this
depends on which (Ring, Controller) you're on.  This is now fixed to look
for a valid configured CCU Spinlock, and fall back to no locking is one
isn't found.