flash_tools/release: add enrollment service scripts to sign step bundle
This is temporary to fix the build process. Ultimately this needs to be
removed as the enrollment service requires network access, which
wouldn't be available with the HSM.
To allow this key injection step needs to be split into a pre-hsm and
hsm part, but this is not trivial so push this intermediate quick fix in
the meantime.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>