Signing Server
* Registration server:

The CGI script for the registration server was slightly refactored to use
common code with the new signing server (HTTP, HSM)

* Signing server:

The signing server is a HTTPS server running on port 4443.
The URL is "https://<nameofserver>:4443/cgi-bin/signing_server.cgi"
It implements the following commands:

** GET requests

** cmd=modulus

Returns the modulus of a key in various formats.
This is similar to the same command for the registration server with
the following changes:

- a "key" argument can be specified to select the proper key. It defaults
to fpk4 like for the registration server if not specified.
- a new value "binary" for the "format" argument  was added in addition
to the ones common with the registration server ("hex", "base64", "c_struct").
Note that "hex" is the default for the registration server and "binary"
is the default for the signing server.
If "binary" is specified, the raw bytes of the modulus for the specified
key is returned in the body of the reply.

** POST requests

The post request is used to sign an image. The request MUST contain the
image to sign and MAY contain a certificate.
Arguments:
- "img": the image to sign
- "cert": the certificate with the public key to use for signing.
The corresponding private key MUST have been imported into the Server HSM.
- "customer_cert": the customer certificate with the public key to use
for signing
- "key": the label of the key to use for signing
- "type": the image type
- "version": the version for the image
- "description": an optional description for the image

An example of building this complex request with Python is in the test script:
 enrollment_tools/tests/test_signing_server.py

Also:

FPK4 backup from f1reg.fungible.com (backup of the encrypted tokens directory)

Well known development only keys FPK2, FPK3, FPK5 imported into the HSM.

enrollment_hsm.py: add a command to import keys -- to be used with the signing server

Python tests for both enrollment server and signing server.

Adjustments to the installation instructions for the signing server.