Move code connected with binary parsing to TraceFileParser.
This change allows us to reuse the various event objects for both
text trace logs and binary trace logs.

The next change will generate the trace events for the text trace logs,
and then will change the code for building relationships between events
so it operates on the events, not dictionaries.