Prompt for the password only once: -create a session only when necessary -cache the session for reuse
The generate_firmware_image.py is imported by higher level scripts to perform the HSM operations. For the SoftHSM, the token and password are hardcoded and the user is never prompted for them. For the real HSMs, the password should be hardcoded but entered interactively by the HSM operator. The generate_firmware_image.py was creating multiple sessions and prompting for the token/password for each of them. The changes in the script preserve the backward compatibility: i.e. the function definitions did not change: the session is created automatically when needed and is cached for reuse so that the user is only prompted once.