Initial support for CSR2 peek
A minimal, hacked up csr peek implementation for CSR2. Includes:

a) CSR2 traversal code from FunHW (csr2api, codegen), both of which have been modified to suit csr peek needs. This part of the code will be thrown away on a future commit which will move this to a newer API.

b) A basic implementation of csr_peek that allows full path queries with "array" support, and also regex search.

c) A global variable in dbgsh to switch between F1 and S1. This is temporary until I can work out a better scheme.

Tested on FS1600 with F1 CSR2 json.