Add s1 i2c to debug client infrastructure
Use a factory method to determine which i2c instance to create. Revamp the s1 i2c interface to match the i2c interface (polymorphism to the rescue!).

At some point we'll want to grab the chip type from the dut.cfg file, but for now it's part of the "connect" arguments.