PEM files as source for certificate modulus
The cert key file used as the source for the modulus of the certificate
can be a PEM file with a Private RSA key, a Public RSA key or a Public
Key. The Private RSA key is usually created with 'openssl genrsa', the
Public RSA Key with 'openssl rsa -RSAPubKey_out' and the Public key with
'openssl rsa -pubout'.