generate_flash.py: Remove the option to specify an enrollment tbs.
The enrollment_tbs was a legacy of the flow on Zync6 where enrollment
certificates were generated on the fly with the local HSM.
Enrollment Certificates should now only be generated by the enrollment server.
It does not make a lot of sense to use an enrollment tbs when the enrollment certificate is well defined and fixed for every serial number.