SWTOOLS-1362: generate endianness-independent HCI definitions
The generator produces both big-endian and little-endian versions of
each structure, cluttering the produced header files. Switch to
producing a single version of each structure that is agnostic of the DPU
endianness. The endianness is resolved later when the headers are
compiled, based on __DPU_BYTEORDER_BIG.

With this the generator no longer concerns itself with the specific
endianness of a CPU, it just distinguishes host CPU vs DPU. Remove
mentions of big vs little endian from the generator and change the
tri-state big_endian parameter in a number of places to the boolean
dpu_endian.

Reduces size of fun_hci.h by about 5K lines.