flash_tools/mmc: do not generate uboot script during final image generation
When a final image containing a signed bootloader script is generated,
avoid generating a new bootloader script file. That file wasn't used
anywhere, but it only makes the output directory confusing, as the
'source' bootloader script then appears to be different than the signed
one that was used for generating the real image.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>