SWOS-6112: Can't dump the CCU till it's been initialized.
If we attempt to dump the Contro Status Register Control Unit before a READ
or WRITE command has happened, it's not initialized and this will cause ECC
Memory faults, etc.  So this commit removes the call to ccu_dump() when a
CONNECT is done and we're in debug mode (-d).  I've added a new "DUMPCCU"
command which can do this if needed.  The new "DUMPCCU" command does not
check for a preceeding READ or WRITE command having been executed because
it's purely for debugging purposes.