dpcsh: generate dummy error response when funOS is not responding
When dpcsh operates in proxy mode there is no response sent to command
socket if the dpcsh cannot open a connection to FunOS. This typically
happens if FunOS has already crashed (or the nvme driver hasn't been
properly initialized, which is a much less likely case) and causes the
python client to hang waiting for response.
Work around this by generating a custom error response that the client
can retrieve to detect the error.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>