nvdimm: add a new wrapper to generate a FunSDK package for nvdimm fw
This is a simple FunSDK package that will allow generating a FunSDK
release from a third-party nvdimm firmware. This is required so that we
have a release-specific nvdimm release available for signing and bundle
generation after adding Fungible signatures.
Special version number (separate from FunSDK version) is stored in the
signed image description and will be used by FunOS to determine at
runtime whether given image needs upgrading or not.
As the version number is part of the third-party version, we cannot rely
on our own versioning scheme, hence this extra indirection.

The extra package that adds custom versions and signature file is
created in order to make sure that the third_party location only
contains an unchanged release from the external firmware provider and
all of our additions/changes are minimized and kept separate.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>