Switch tuple to use datetime object.
Instead of converting back and forth, keep everything as a datetime object. We continue to pass the usec value around as the second entry in the tuple. That's because I'm still worried about the loss of precision in the FunOS case where we have to convert into a double-precision floating point number.