SWOS-12155: generate HCI files for a specific endianness
Two new codegen options are added, 'be' and 'le', that work alongside
the existing 'swap'. swap on its own produces files that handle both
endiannesses, the choice is made at the time those files are compiled
through CPP defines (current behavior). swap,be produces BE files, the
choice is made at the time the files are generated. Likewise for
swap,le.