Add support for endian-specifying types.
Generator will use the Linux-style types with endianness - __le16, __be32, etc.
Types with endianness won't be swapped when reading or writing the structures.

Types with endianness are allowed in fields of natural type as well as
bitfields.

FunOS will need the Linux endianness types added before this feature can be
used.