firmware image: add chip type into header
We have 32 spare bytes in the attributes field. Use 3 bytes to indicate
chip type for the given image. This will allow authentication/upgrade
code to reject firmware that isn't correct for the given DPU type.
Value '0' is left for compatibility with existing images/systems and
will be used as an indicator that the firmware does not require
validation.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>