flash_tools/setup.sh: properly skip eepr upgrade when unsupported
Old FunOS builds do not report sku type via the required dpc query, so
the upgrade fails as it attempts to upgrade 'null' eeprom image.
Explicitly skip upgrade not only on dpcsh errors, but also when 'null'
is returned, as this is never expected to be a valid string identifier
for the eepr type.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>