Handle trace overflow.
Cache miss tracing is overflowing the trace buffers. The correct solution according to MIPS spec would be to detect the infamous TMOAS-TPC pair after an overflow (TF5) and restart from that. However, the hardware seems to be ignoring the spec when cache miss filtering is enabled.

So we use a heuristic where we look for the resumption of TPC-TLA/TSA pairs. This seems to work well enough in practice.