release.py: do not include emmc.bin in the bundle
We have switched to upgrade using mmc1.bin a long time ago and emmc.bin
has only been kept for compatibility with ancient FunOS builds. It is
not expected to be useful anymore, so stop including this file in the
bundle especially that it uses a lot of RAM when unpacked before
installation.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>