Work around missing support for thread-locals.
gdb throws exceptions when working with thread-local symbols. The workaround is to identify these thread-local symbols via objdump, and append them to the master list.

This is the first step. The next step is (possibly) to mock the gdb types better so we can walk into per-vp structs and arrays.