Memory map the input hex dump files for HBM dumps.
Previous code would do reads of multiple lines, and had expectations
about exactly how many bytes back it would get on the read.  This worked
on previous chips but for whatever reason it started failing with S2
traces, perhaps because of the size.

Instead, we'll memory map all the files, and just follow pointers to
process all the text hex dump lines and convert them to binary for later
unsharding.

Tested on saved traces and works, in process of testing on Palladium
to make sure the new scheme works ok in the acutal pipeline.