cfggen: allow non-compliant json in fungible boards
I want to add longer board descriptions in the fungible boards json.
Embedding long strings in json isn't very nice for reading (and they are
generally meant to be descriptive and not necessarily useful for
anything else other than help string), so allow a non-compliant json so
that we can embed newlines in json values.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>