cfggen: support multiple asic types per board/skuid
As there can be potentially different sets of extra flags added to eepr
files, the majority of the change is to support generation of separate
skuid eepr files ({chip}_{skuid}).
In theory if it were not for the flags, we could have the same skuid
file used, although they need to be signed separately for each chip
anyway, so in the end we'd need to differentiate them anyway.

At the moment some files end up with silly names, as we'll have
f1_f1_dev_board or s1_s1_dev_board_x16_host_upsw_4x4_dnsw because we
have the chip name embedded in sku name. Well, the easiest way to deal
with this is to ignore that ;)

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>