Handle pyelftools note type change.
Before v0.28 of pyelftools, the note type was a string (decoded from the raw bytes using latin-1). In order to read the note as a struct I had to reencode the string into raw bytes.

Apparently a pyelftools contributor realized strings for note types was a bad idea and changed the type to bytes....

New workaround: check the type and do the right thing so various versions of pyelftools will continue to work.