cfggen: fix emulation eeprom generation
Fix breakage introduced by 8f9f56e4 which resulted in lack of sku
eeproms. This is due to the fact that get_chips_for_sku_id only returns
a valid chip for board SKUs. In case of emu targets, it returns an empty
set (None prior to 8f9f56e4) and as a result we did not iterate over the
list of targets to generate.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>