upgrade: try to sync b-persist config in the background
When the firmware upgrade completes, it is up to the user to perform a
system reset, which can happen a lot later than immediately after
installation. If user configuration changes after sync has been done and
before the reset, those changes would not be reflected after a reboot.

Add a daemon that will try to sync the configs periodically to reduce
probability of such issues.
The code is now duplicated (with only minor changes) between b-persist
and bundle installer. Ideally it should exist in b-persist only so that
the rootfs can determine what to do, but in such cases the change would
only be reflected on systems that have already been updated with this
change. Duplicating the logic ensures that we can do the background sync
even on existing systems.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>