Retention fixer-upper.
Set retention tags on binaries and metadata that do not have tags. The retention time is determined from the original build path in the metadata (that code is cribbed from an existing cleanup script).

Includes a dry-run mode. Also includes a way to continue from a nearby point in case the script fails midway.