upgrade: (maybe) fix b-persist sync issue
It's been reported that after a downgrade, _sometimes_ the firmware
version verification fails due to a mismatch between expected and
running firmware versions.
The problem is that the _correct_ firmware is running, and the expected
version (as stored in /persist) is wrong. It seems that the process IDs
assigned by platform agent are not-so-random and can often repeat and in
some cases we could have a stale version file.
I suspect that after recent b-persist sync changes, we could end up
doing rsync without a filesystem sync afterwards, so if the system is
rebooted quickly, the changes might not be reflected in the filesystem.
That's at least my working theory right now ... let's see if explicit
syncs resolve that.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>