Fix parsing HostUnit and HostUnitController errors in pcicfgutil
There were a number of errors in the parsing of the "hu_{ring}" and
"ctl_{cid}" elements of HostUnit and HostUnitController.

First, we were trying to do an atoi() on the "_{ring}" and "_{cid}" keys,
forgetting to skip past the leading "_".

Second, for HostUnitController, our inner JSON iterator loop was referencing
the HostUnitController JSON instead of the "hu_{ring}" sub-JSON.  The naming
was super confusing and things were really deeply indented making it hard to
understand, so I reformatted all of that.