i2c_chal: (socket) fix invalid read request length
Read request can be set to a maximum of 64 bytes, as there are only 6
bits available in the request field. Current code would set the number
to 65, which from device PoV indicates a read of 1 bytes.
This somehow works correctly on HW, probably because the number is not
relevant to the HW block, as the reads are then driven by the i2c master
clock.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>