Remove the PCIe Link Configuration from the NOR FLASH EEPROM Image
For the nonce, we're no longer focusing on PCIe Adapters (likely to be
changed, but that's the future).  In the mean time, our initial choice to
put the PCIe Link Configuration (struct hw_hsu_api_link_config) in the
NOR FLASH EEPROM Image ("eepr") file was a mistake.  This unfortunately
bound the SKU ID to a particular PCIe Configuration.  When/if we bring back
the Early PCIe Initialization in SBPFirmware, we'll want to read that from a
completely separate NOR FLASH file in order to untangle these concepts.