release.py: accept `--funos-type` argument to select funos variant
It is a free text entry field with no restricted set of values so that
we don't need to modify the script any time a new build variant is
created. If a wrong value is passed then the script will fail - bad luck.
If no value is specified then the script will behave as usual and use
funos-{chip}.elf.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>