fwupgrade: prepare for decoupled pufr/frmw
In preparation for enabling decoupled SBP firmware and pufrom, change
the installation logic to:
1) never upgrade PUFR by default, it will still be possible to perform
PUFR upgrade by specifying that explicitly to run_fwupgrade script (note
that at least right now it's not going to work with standard bundles,
requires manual unpacking of bundle and running run_fwupgrade with pufr
arg explicitly)
2) always (if needed) attempt to upgrade both FRMW and SBPF images. On
current firmwares this will result in FRMW upgrade error (which will be
ignored) and followed by SBPFW installation. If FRMW succeeds, then SBPF
upgrade will not be attempted. That way upgrades of either type of the
system can be supported in the transition period.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>