fwupgrade: remove support for HCI-based interface
This interface has not been used for a long time now and doesn't work
with latest core funos at all. All other tools and funos code are about
to be removed as well.
This greatly simplifies the flow of the upgrade script logic as well as
allows getting rid of a few no longer used cmdline args.

While testing the cleaned up code I've discovered a few minor bugs in
dry-run handling for dpc upgrade, so these are fixed as well.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@fungible.com>