Add API to allow deserialization of streams.
Need some way to identify how many bytes need to be read from the stream, especially for blocking sockets. The new code is a direct copy from the FunOS implementation.

Why the duplication? Because I need to deserialize fun_json blobs in Windows, and the SDK funclient library does not support Windows. I've started looking at how to compile funclient on Windows, but until I get that working this is my stopgap approach.