run_fwupgrade: fix upgrade error with `-u` option
When using `-u fourcc` *and* a build with the same version number as
set via `--version` is found, then we incorrectly stop processing the
whole directory instead of skipping this entry.

This results in an error when eg. trying to upgrade inactive copy of
bcfg to a bundle number as the one that is currently active resulting in
a system where all images except bcfg are upgraded properly to have
matching versions.

Signed-off-by: Marcin Nowakowski <mnowakowski@microsoft.com>