release.py: include extra dpc scripts in xdata for mfginstallers
In order to allow extra commands to be executed by the installation
process using mfginstaller, extend the release script to allow embedding
custom dpc scripts in xdata.

Rather than expecting the mfg installer process on the BMC to run a set
of arbitrary dpc commands that may be needed, the intention is to allow
a level of flexibility allowing the BMC to specify a single script to
invoke a set of expected actions and let the script define the actual
funos actions that need to be taken.

Signed-off-by: Marcin Nowakowski <mnowakowski@microsoft.com>