swos-22948: i2c_chal: tentative fix for qemu flaky test failure
Recent changes to i2c_chal have added a random failure in sbp-qemu when
sending a challenge interface response (assert in qemu challenge
response).

It's quite random and not reliably reproducible, but I suspect that the
recently introduced sleep results in the response being queued in qemu's
character device when SBP generates a response which happens before the
read request from client library arrives. The result is that the
response is queued in the chardev and SBP assumes it was already sent,
but then the request arrives for that same response to be read from the
challenge interface buffer.

Potentially the qemu side of the implementation needs some improvements
to handle this scenario better, but given that we had not seen the failure
previously - it seems to work quite reliably so before we get a more
correct fix, back out the extra delay that isn't really necessary and is
potentially fatal in case of qemu ...

Signed-off-by: Marcin Nowakowski <mnowakowski@microsoft.com>