generator: Fix be and le codegen options.
While the generator can target HCI with specific endianness through the
'le' and 'be' options, the structures it generates still use __dpuxx
types and rely on preprocessor defines. Adjust the generator so it can
use the proper __[bl]exx types directly in the structures it generates.

Update the Linux header template to omit __dpuxx types for LE and BE
HCI.