generator: Rework name mangling
The previous mangling implementation mangled names early, before adding
them to symbol tables. This meant that uses of field names in macros,
function names, function parameters, and local variables were all using
the mangled names. This was wrong in the case of macro and function
names.

Rework the mangling implementation so mangling is done later, at the
time names are emitted. Both mangled and unmangled names are available
to each user and the choice between them is left to them.