bmc_sbp_chal: add delay after inject certificate
When certificate is injected while in bootrom, it takes quite a
significant time for the response to be returned from SBP to the
challenge interface.

While in bootrom all operations are slower than normal runtime (no PLLs)
and we potentially need to do some of the crypto init as well if it's
started for the first time.

It has been observed that issuing an unlock command for the first time
fails due to lack of response from DPU's i2c, but typically works on
second invocation (because any one-time initializations have completed
by then).

Signed-off-by: Marcin Nowakowski <mnowakowski@microsoft.com>