bmc_sbp_chal: fix C library loading on fs1600pro/ds400 BMCs
LoadLibrary does not return any meaningful error nor does allow checking
for successful loading of the library if an invalid path is passed.
find_library relies on ldconfig and/or ld being present in the rootfs,
but that's not the case for our BMCs, so it never returns a valid
library to load. Use default lib locations as a fallback for that case.

Signed-off-by: Marcin Nowakowski <mnowakowski@microsoft.com>