release.py: add a trivial emmc contents check
Some recent FunStorage builds contain invalid emmc images where the MBR
location is empty. I don't see how this could happen, as if the tools to
generate the image fail, it should raise a process error and abort the
build. It appears that the issue is now resolved, but a few builds have
been generated incorrectly (no changes to sdk were made for storage
builds during that time, so it cannot be attributed to any recent
changes here).

I have not seen jenkins logs for the builds that have invalid images, so
I don't know if anything has been reported in the logs either (if I get
access to them, I might push more fixes).

As an extra precaution add an extra check to verify that we have an MBR
flag at bytes 510-511 and fail the build if it's missing.

Signed-off-by: Marcin Nowakowski <mnowakowski@microsoft.com>