Make perf processing independent of fun-on-demand.
It was a mistake in 2019 to couple perf decoding with the fun-on-demand infrastructure. The files in this PR have been copied from the PalladiumOnDemand repo and modified to avoid usage of fun-on-demand concepts and infrastructure.

The next step is to remove the PalladiumOnDemand copies and modify fun-on-demand to invoke these scripts.